requirements.txt
asciimatics==1.13.0
certifi==2021.10.8
colorama==0.4.4
crayons==0.4.0
DateTime==4.3
elasticsearch==7.10.0
emoji-country-flag==1.3.0
future==0.18.2
Pillow==8.4.0
pyfiglet==0.8.post1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
urllib3==1.26.7
utm==0.7.0
wcwidth==0.2.5
zope.interface==5.4.0
